requirements.txt
aniso8601==9.0.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
config==0.5.1
Flask==3.0.2
Flask-Mail==0.9.1
Flask-RESTful==0.3.10
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
PyMySQL==1.1.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2024.1
six==1.16.0
Werkzeug==3.0.1
